Feature/common

+++ Edulingual.Domain/Entities/Course.cs
@@ -8,7 +8,7 @@ public class Course : BaseEntity
     public string Title { get; set; } = string.Empty;
     public string Description { get; set; } = string.Empty;
     public string Duration { get; set; } = string.Empty;
-    public decimal Fee { get; set; }
+    public double Fee { get; set; }
     public Guid OwnerId { get; set; }
     public Guid CourseAreaId { get; set; }
     public Guid CourseLanguageId { get; set; }
